+++ src/main.py
@@ -1,7 +1,8 @@
 from fastapi import FastAPI
-from .models import *
 from .routers.users import router as user_router
 from .routers.sessions import router as session_router
+from .routers.screen import router as screen_router
+from .routers.event import router as event_router
 from .routers.user_screen_time import router as user_screen_time_router
 from .routers.screen_time_stats import router as screen_time_stats_router
 from .routers.avg_engmt_time import router as avg_engmt_time_router
@@ -16,6 +17,8 @@ app = FastAPI(
 )
 app.include_router(user_router)
 app.include_router(session_router)
+app.include_router(screen_router)
+app.include_router(event_router)
 app.include_router(user_screen_time_router)
 app.include_router(screen_time_stats_router)
 app.include_router(avg_engmt_time_router)
