+++ tests/test_event.py
@@ -1,18 +1,11 @@
 import uuid
-
 from fastapi.testclient import TestClient
-from src.user import delete_user_from_database
-from src.session import *
-from src.event import *
+from utils.utils import delete_user_from_database, deleteEvent
 import json
-from src.session import app as session_app
-from src.user import app as user_app
-from src.event import app as event_app
+from src.main import app
 
 
-session_client = TestClient(session_app)
-user_client = TestClient(user_app)
-event_client = TestClient(event_app)
+client = TestClient(app)
 
 
 """
@@ -26,7 +19,7 @@ def test_add_event_successfully():
     new_session_id = uuid.uuid4()
     try:
         payload = {"userID": str(new_user_id)}
-        response = user_client.post("/api/analysis/add-user/", json=payload)
+        response = client.post("/api/analysis/add-user/", json=payload)
         add_user_response = json.loads(response.content.decode('utf-8'))
         assert add_user_response['status code'] == 200
         assert add_user_response['message'] == "Added user successfully"
@@ -35,7 +28,7 @@ def test_add_event_successfully():
                            "startTime": "2023-04-19 12:30:45"
                            }
 
-        session_response = session_client.post("/api/analysis/record-session-start-time/", json=session_payload)
+        session_response = client.post("/api/analysis/record-session-start-time/", json=session_payload)
         assert session_response.status_code == 200
         add_session_response = json.loads(session_response.content.decode('utf-8'))
         assert add_session_response['status code'] == 200
@@ -47,7 +40,7 @@ def test_add_event_successfully():
             "eventName": "scroll",
             "occurTime": "2023-04-19 12:30:45"
             }
-        add_event_response = event_client.post("/api/analysis/record-event/", json=event_payload)
+        add_event_response = client.post("/api/analysis/record-event/", json=event_payload)
         assert add_event_response.status_code == 200
         add_event_response = json.loads(add_event_response.content.decode('utf-8'))
         assert add_event_response['status code'] == 200
@@ -65,13 +58,13 @@ def test_missing_sessionID():
     new_session_id = uuid.uuid4()
     try:
         payload = {"userID": str(new_user_id)}
-        response = user_client.post("/api/analysis/add-user/", json=payload)
+        response = client.post("/api/analysis/add-user/", json=payload)
         assert response.status_code == 200
         session_payload = {"userID": str(new_session_id),
                            "sessionID": str(new_session_id),
                            "startTime": "2023-04-19 12:30:45"
                            }
-        session_response = session_client.post("/api/analysis/record-session-start-time/", json=session_payload)
+        session_response = client.post("/api/analysis/record-session-start-time/", json=session_payload)
         assert session_response.status_code == 200
 
         #add event that is based on this session
@@ -80,7 +73,7 @@ def test_missing_sessionID():
             "eventName": "scroll",
             "occurTime": "2023-04-19 12:30:45"
             }
-        add_event_response = event_client.post("/api/analysis/record-event/", json=event_payload)
+        add_event_response = client.post("/api/analysis/record-event/", json=event_payload)
         add_event_response = json.loads(add_event_response.content.decode('utf-8'))
         assert add_event_response['status code'] == 400
         assert add_event_response['message'] == "missing required field: sessionID"
@@ -94,13 +87,13 @@ def test_missing_eventName():
     new_session_id = uuid.uuid4()
     try:
         payload = {"userID": str(new_user_id)}
-        response = user_client.post("/api/analysis/add-user/", json=payload)
+        response = client.post("/api/analysis/add-user/", json=payload)
         assert response.status_code == 200
         session_payload = {"userID": str(new_session_id),
                            "sessionID": str(new_session_id),
                            "startTime": "2023-04-19 12:30:45"
                            }
-        session_response = session_client.post("/api/analysis/record-session-start-time/", json=session_payload)
+        session_response = client.post("/api/analysis/record-session-start-time/", json=session_payload)
         assert session_response.status_code == 200
 
         #add event that is based on this session
@@ -109,7 +102,7 @@ def test_missing_eventName():
             "eventName": "",
             "occurTime": "2023-04-19 12:30:45"
             }
-        add_event_response = event_client.post("/api/analysis/record-event/", json=event_payload)
+        add_event_response = client.post("/api/analysis/record-event/", json=event_payload)
         add_event_response = json.loads(add_event_response.content.decode('utf-8'))
         assert add_event_response['status code'] == 400
         assert add_event_response['message'] == "missing required field: eventName"
@@ -122,13 +115,13 @@ def test_missing_occurTime():
     new_session_id = uuid.uuid4()
     try:
         payload = {"userID": str(new_user_id)}
-        response = user_client.post("/api/analysis/add-user/", json=payload)
+        response = client.post("/api/analysis/add-user/", json=payload)
         assert response.status_code == 200
         session_payload = {"userID": str(new_session_id),
                            "sessionID": str(new_session_id),
                            "startTime": "2023-04-19 12:30:45"
                            }
-        session_response = session_client.post("/api/analysis/record-session-start-time/", json=session_payload)
+        session_response = client.post("/api/analysis/record-session-start-time/", json=session_payload)
         assert session_response.status_code == 200
 
         #add event that is based on this session
@@ -137,7 +130,7 @@ def test_missing_occurTime():
             "eventName": "scroll",
             "occurTime": ""
             }
-        add_event_response = event_client.post("/api/analysis/record-event/", json=event_payload)
+        add_event_response = client.post("/api/analysis/record-event/", json=event_payload)
         add_event_response = json.loads(add_event_response.content.decode('utf-8'))
         assert add_event_response['status code'] == 400
         assert add_event_response['message'] == "missing required field: occurTime"
@@ -149,7 +142,7 @@ def test_invalid_sessionID():
     new_user_id = uuid.uuid4()
     try:
         payload = {"userID": str(new_user_id)}
-        response = user_client.post("/api/analysis/add-user/", json=payload)
+        response = client.post("/api/analysis/add-user/", json=payload)
         assert response.status_code == 200
 
         #add event that is based on this session
@@ -158,7 +151,7 @@ def test_invalid_sessionID():
             "eventName": "scroll",
             "occurTime": "2023-04-19 12:30:45"
             }
-        add_event_response = event_client.post("/api/analysis/record-event/", json=event_payload)
+        add_event_response = client.post("/api/analysis/record-event/", json=event_payload)
         add_event_response = json.loads(add_event_response.content.decode('utf-8'))
         assert add_event_response['status code'] == 400
         assert add_event_response['message'] == "SessionID does not exists"
@@ -172,13 +165,13 @@ def test_invalid_time():
     new_session_id = uuid.uuid4()
     try:
         payload = {"userID": str(new_user_id)}
-        response = user_client.post("/api/analysis/add-user/", json=payload)
+        response = client.post("/api/analysis/add-user/", json=payload)
         assert response.status_code == 200
         session_payload = {"userID": str(new_session_id),
                            "sessionID": str(new_session_id),
                            "startTime": "2023-04-19 12:30:45"
                            }
-        session_response = session_client.post("/api/analysis/record-session-start-time/", json=session_payload)
+        session_response = client.post("/api/analysis/record-session-start-time/", json=session_payload)
         assert session_response.status_code == 200
 
         #add event that is based on this session
@@ -187,7 +180,7 @@ def test_invalid_time():
             "eventName": "scroll",
             "occurTime": "2023-04-19"
             }
-        add_event_response = event_client.post("/api/analysis/record-event/", json=event_payload)
+        add_event_response = client.post("/api/analysis/record-event/", json=event_payload)
         add_event_response = json.loads(add_event_response.content.decode('utf-8'))
         assert add_event_response['status code'] == 400
         assert add_event_response['message'] == "provided datetime is either not in mysql datetime format nor an invalid datetime"
